# WareHouseProjects/api/docs/swagger.yaml
definitions:
...
  models.ComingTable:
    properties:
      branch_id:
        type: string
      coming_id:
        type: string
      created_at:
        type: string
      date_time:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/models.TableType'
      updated_at:
        type: string
    type: object
...
  models.CreateComingTable:
    properties:
      branch_id:
        type: string
      coming_id:
        type: string
      date_time:
        type: string
    type: object
...
  models.GetAllComingTableRequest:
    properties:
      branch_id:
        type: string
      coming_id:
        type: string
      limit:
        type: integer
      page:
        type: integer
    type: object
...
  models.TableType:
    enum:
    - finishied
    - in_process
    type: string
    x-enum-varnames:
    - Finishied
    - InProcess
...
  models.UpdateComingTable:
    properties:
      branch_id:
        type: string
      coming_id:
        type: string
      date_time:
        type: string
      id:
        type: string
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /coming_table:
    get:
      consumes:
      - application/json
      description: gets all Coming_Table based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: coming_id
        in: query
        name: coming_id
        type: string
      - description: branch_id
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllComingTableRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: LIST Coming_Table
      tags:
      - coming_table
    post:
      consumes:
      - application/json
      description: add ComingTable data to db based on given info in body
      parameters:
      - description: ComingTable data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateComingTable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATE ComingTable
      tags:
      - coming_table
  /coming_table/{id}:
    delete:
      consumes:
      - application/json
      description: deletes ComingTable by id
      parameters:
      - description: id of ComingTable
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE ComingTable BY ID
      tags:
      - coming_table
    get:
      consumes:
      - application/json
      description: gets ComingTable by ID
      parameters:
      - description: ComingTable ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ComingTable'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - coming_table
    put:
      consumes:
      - application/json
      description: UPDATES COMINGTABLE BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of ComingTable
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: ComingTable data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateComingTable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE COMINGTABLE
      tags:
      - coming_table
...
swagger: "2.0"